requirements.txt
qdrant-client
sentence-transformers
anthropic
python-dotenv
pydantic
